completed

+++ Backend/controller/cartController.js
@@ -9,9 +9,7 @@ exports.get = async function (req, res) {
             where: { user_id: parseInt(user_id) },
             select: {id: true, item_type: true, url_slug: true, quantity: true, price: true, total_amount: true }
         });
-        if (cart.length === 0) {
-            return res.status(404).json({ success: false, message: 'Cart is Empty' });
-        }
+        
         res.status(200).json({ success: true, cart });
     } catch (error) {
         console.error(error);
@@ -44,17 +42,33 @@ exports.add = async function (req, res) {
         if (!searchItem){
             return res.status(404).json({ success: false, message: `Item doesn't exists` });
         }
-        if (searchItem.quantity < 1){
+        if (searchItem.quantity < quantity){
             return res.status(404).json({ success: false, message: 'Out of stock' });  // which item out of stock show them
         }
-        let total_amount = parseInt(searchItem.price) * parseInt(quantity)
 
-        // add new item in cart
-        const newCartItem = await prisma.cart.create({
-            data: { user_id: parseInt(user_id), item_type, url_slug, quantity: parseInt(quantity), price : searchItem.price, total_amount}
-        });
-        // res.status(200).json({ success: true, message: 'Item added to Cart' });
-        res.status(200).json({ success: true, newCartItem, message: 'Item added to Cart' });
+        // check if already exists
+        const checkCartItem = await prisma.cart.findFirst({
+            where: { url_slug }
+        })
+        let total_amount;
+
+        if(checkCartItem){
+            total_amount = parseInt(searchItem.price) * (parseInt(quantity) + checkCartItem.quantity )
+            // update item in cart
+            const updatedcartItem = await prisma.cart.update({
+                where: { id: checkCartItem.id },
+                data: { quantity: (parseInt(quantity) + checkCartItem.quantity ), price : searchItem.price, total_amount}
+            });
+            res.status(200).json({ success: true, updatedcartItem, message: 'Item added to Cart' });
+        }
+        else{
+            total_amount = parseInt(searchItem.price) * parseInt(quantity)
+            // add new item in cart
+            const updatedcartItem = await prisma.cart.create({
+                data: { user_id: parseInt(user_id), item_type, url_slug, quantity: parseInt(quantity), price : searchItem.price, total_amount}
+            });
+            res.status(200).json({ success: true, updatedcartItem, message: 'Item added to Cart' });
+        }
         
     } catch (error) {
         console.error(error);
@@ -66,11 +80,14 @@ exports.add = async function (req, res) {
 exports.remove = async function (req, res) {
     try {
         const user_id = req.user.id
-        const { url_slug } = req.body;
+        const url_slug = req.params.url_slug;
         
         const itemFindToDelete = await prisma.cart.findFirst({
             where: { user_id, url_slug }
         });
+        if(!itemFindToDelete){
+            return res.status(404).json({ success: false, message: 'Item not found in cart' });
+        }
 
         const deletedCartItem = await prisma.cart.delete({
             where: { id: itemFindToDelete.id }
@@ -80,7 +97,7 @@ exports.remove = async function (req, res) {
             return res.status(404).json({ success: false, message: 'Cart is empty' });
         }
         
-        res.status(200).json({ success: true, message: 'Item removed from cart' });
+        res.status(200).json({ success: true, url_slug, message: 'Item removed from cart' });
         // res.status(200).json({ success: true, removedItem: deletedCartItem, message: 'Item removed from cart' });
         
     } catch (error) {
@@ -93,34 +110,49 @@ exports.remove = async function (req, res) {
 exports.update = async function (req, res) {
     try {
         const user_id = req.user.id;
-        const { url_slug, quantity } = req.body;
+        const { item_type, url_slug, inc } = req.body;
 
         // Search for that item
         const searchItem = await prisma.cart.findFirst({
             where: { user_id: parseInt(user_id), url_slug }
         });
         if (!searchItem){
-            return res.status(404).json({ success: false, message: 'Item not found in cart' });
-        }
-        if (searchItem.quantity < 1){
-            return res.status(404).json({ success: false, message: 'Item out of stock' });
+            return res.status(404).json({ success: false, message: 'Item not in cart' });
         }
 
-        let updatedCartItem;
-        if (parseInt(quantity) === 0) {
-            await prisma.cart.delete({
-                where: { id: searchItem.id }
-            })
-            return res.status(200).json({ success: true, message: 'Item removed from cart' });
+        const newQuantity = inc ? searchItem.quantity + 1 : searchItem.quantity - 1;
+        const newTotalAmount = inc ? searchItem.total_amount + searchItem.price : searchItem.total_amount - searchItem.price;
+
+        // Select table (productDetails or accessory)
+        let table;
+        if (item_type === "Product"){
+            table = "productDetail"
+        }
+        else if(item_type === "Accessory"){
+            table = "accessory"
         }
         else{
-            updatedCartItem = await prisma.cart.update({
-                where: { id: searchItem.id },
-                data: { quantity: parseInt(quantity) }
-            });
+            return res.status(404).json({ success: false, message: 'Invalid item type' });
         }
 
-        res.status(200).json({ success: true, updatedCartItem, message: 'Cart updated' });
+        // Find that item
+        const checkItem = await prisma[table].findFirst({
+            where: { url_slug }
+        })
+        if (!checkItem){
+            return res.status(404).json({ success: false, message: `Item doesn't exists` });
+        }
+        if (checkItem.quantity < newQuantity){
+            return res.status(404).json({ success: false, message: 'Out of stock' }); 
+        }
+        
+        // update values
+        await prisma.cart.update({
+            where: { id: searchItem.id },
+            data: { quantity: newQuantity, total_amount: newTotalAmount }
+        })
+
+        res.status(200).json({ success: true, newTotalAmount, url_slug, inc});
         
     } catch (error) {
         console.error(error);
